Moved SalesOfferPackageInfo type into salesOfferPackages.ts
