+++ src/pages/api/salesOfferPackages.ts
@@ -5,7 +5,12 @@ import { updateSessionAttribute } from '../../utils/sessions';
 import { ErrorInfo, NextApiRequestWithSession } from '../../interfaces';
 import { SOP_INFO_ATTRIBUTE } from '../../constants';
 import { purchaseLocationsList, paymentMethodsList, ticketFormatsList } from '../salesOfferPackages';
-import { SalesOfferPackageInfo } from '../describeSalesOfferPackage';
+
+export interface SalesOfferPackageInfo {
+    purchaseLocations: string[];
+    paymentMethods: string[];
+    ticketFormats: string[];
+}
 
 export interface SalesOfferPackageInfoWithErrors extends SalesOfferPackageInfo {
     errors: ErrorInfo[];
